modified:   Spotify/Main.java Validacion año Album validacion duracion Cancion validacion seleccion de menu validacion seleccion de album para nueva cancion validacion seleccion de album para ver album validacion del menu principal validacion del menu buscar albumes por año validar todos los ingresos de datos para no permitir datos vacios
modified:   Spotify/Cancion.java
cambio de tipo de dato de duracion de int a String

+++ Spotify/Cancion.java
@@ -1,9 +1,9 @@
 public class Cancion {
     private String titulo;
-    private int duracion;
+    private float duracion;
 
     // Constructor
-    public Cancion(String titulo, int duracion) {
+    public Cancion(String titulo, float duracion) {
         this.titulo = titulo;
         this.duracion = duracion;
     }
@@ -13,7 +13,7 @@ public class Cancion {
         return titulo;
     }
 
-    public int getDuracion() {
+    public float getDuracion() {
         return duracion;
     }
 }
